requirements.txt
Flask==3.1.2
python-dotenv==1.0.1
replicate==1.0.7
docling==2.64.1